+++ Source/CS/Desktop/Model/Dto/MesProOutDetailDto.cs
@@ -41,7 +41,7 @@ namespace Model.Dto
         /// <summary>
         /// 数量
         /// </summary>
-        public Double P_Qty{ set; get; }
+        public decimal P_Qty{ set; get; }
 
         /// <summary>
         /// 批次
