+++ Source/CS/Desktop/Model/Dto/MesWeightRecordDto.cs
@@ -46,7 +46,7 @@ namespace Model.Dto
         /// <summary>
         /// 数量
         /// </summary>
-        public Double W_Qty{ set; get; }
+        public decimal W_Qty{ set; get; }
 
         /// <summary>
         /// 批次
