+++ Source/CS/Desktop/Model/System/MesCollarDetailEntity.cs
@@ -35,7 +35,7 @@ namespace Model
         /// <summary>
         /// 数量
         /// </summary>
-        public Double C_Qty{ set; get; }
+        public decimal C_Qty{ set; get; }
         /// <summary>
         /// 批次
         /// </summary>
