+++ Source/CS/Desktop/Model/System/MesInventoryEntity.cs
@@ -35,7 +35,7 @@ namespace Model
         /// <summary>
         /// 数量
         /// </summary>
-        public Double I_Qty{ set; get; }
+        public decimal I_Qty{ set; get; }
         /// <summary>
         /// 批次批次以入库时间（yyyymmdd)
         /// </summary>
