+++ Source/CS/Desktop/Model/System/MesInventoryLSEntity.cs
@@ -39,7 +39,7 @@ namespace Model
         /// <summary>
         /// 数量
         /// </summary>
-        public Double I_Qty{ set; get; }
+        public decimal I_Qty{ set; get; }
         /// <summary>
         /// 批次
         /// </summary>
