+++ Source/CS/Desktop/Model/System/Mes_OutWorkShopDetailEntity.cs
@@ -14,9 +14,9 @@ namespace Model
         public string O_GoodsCode { set; get; }
         public string O_GoodsName { set; get; }
         public string O_Unit { set; get; }
-        public Double O_Qty { set; get; }
+        public decimal O_Qty { set; get; }
         public string O_Batch { set; get; }
         public string O_Remark { set; get; }
-        public Double O_Price { set; get; }
+        public decimal O_Price { set; get; }
     }
 }
